initial commit for Class tracker created for purposes of: -Final Project -CPRE 308 -2015 @ ISU
This project was intended to mimic the behalvior of a course tracking system.
The project is incomplete, but is curently in a functional beta release.
The user can register classes, class info and create a printout of current records.
The server acts as a central tracking system making use of IPC.
The client communicates with the server using shared mem struct cclass_shm.h

Please use care when modify code and looping structure.
Utilizes flagging in shared memory to indicate creation of new job between Client and Server.