initial commit for MTCP(Multipath-Transmission-Control-Protocal) created for purpose of: -Final Project -CPRE 489 -2016 @ ISU
This set of server-client partner programs are intended to recreate the behalvior of an multi threading, multi path transmission algorythm.
The server-client pair establish connections over four alternate ports, 1 control and three data transfer.
A data stream is parsed and transfer in sequences across the 3 data ports using the data connection to manage transfer flow.
Use of threaded child processes provide a proposed increase in transfer rate by utilizing a sending/recieving buffer for decoupling/reassembling data stream.

This client-server pair is for demonstrational purposes only. it is designed only as proof of concept and may not perform exactly as intended. It is not complete and is not in alpha release at this time.

Please enjoy...