Login AngularJS crypto
This project was created for demonstrational purposes only. The code was
ment to test a proof of concept and satisfy course requirments for ComS
final project.

It may not perform as expected under current build.

Modifications must be made to dbConnector.php to allow connection to sql
database.

Sql build file contains a scrypt for creating the needed database.

This is a demonstration of AngularJS, PHP and crypto algorythms.