Video Chat Client and Server
- ComS 319
- 2016 @ ISU

The project was created in reponse to a project assignment requiring the
use of swing and application of java frameworks.

This project alows multiple users to send chat mesages and camera
captures to other users logged into the server.

Server must be started before attempting to connect with client.

Additional drivers and java api will be required to  opperate the
cli-server correctly.